requirements.txt
apscheduler==3.6.3; python_version >= "3.6"
attrs==21.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
autobahn==22.5.1; python_version >= "3.7"
automat==20.2.0; python_full_version >= "3.6.7" and python_version >= "3.6"
binance-connector==1.13.0; python_version >= "3.6"
cachetools==4.2.2; python_version >= "3.6" and python_version < "4.0"
certifi==2022.5.18.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0; python_version >= "3.7"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.5.0"
constantly==15.1.0; python_full_version >= "3.6.7" and python_version >= "3.6"
cryptography==37.0.2; python_version >= "3.7"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
hyperlink==21.0.0; python_full_version >= "3.6.7" and python_version >= "3.7"
idna==3.3; python_full_version >= "3.6.7" and python_version >= "3.7"
incremental==21.3.0; python_full_version >= "3.6.7" and python_version >= "3.6"
loguru==0.6.0; python_version >= "3.5"
pyasn1-modules==0.2.8; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
pydantic==1.9.1; python_full_version >= "3.6.1"
pyopenssl==22.0.0; python_version >= "3.6"
python-dotenv==0.20.0; python_version >= "3.5"
python-telegram-bot==13.12; python_version >= "3.6"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pytz==2022.1; python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
service-identity==21.1.0; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.7" and python_version >= "3.6"
sqlalchemy==1.4.37; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
tornado==6.1; python_version >= "3.6"
twisted-iocpsupport==1.0.2; python_full_version >= "3.6.7" and python_version >= "3.6" and platform_system == "Windows"
twisted==22.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
txaio==22.2.1; python_version >= "3.7"
typing-extensions==4.2.0; python_version >= "3.7" and python_full_version >= "3.6.7"
tzdata==2022.1; platform_system == "Windows" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
tzlocal==4.2; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
zope.interface==5.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
